move subscriber to its own module
